requirements.txt
watchdog==4.0.0
mistralai==0.1.3
python-dotenv==1.0.1
ftfy==6.1.3
PyYAML==6.0.1
pydantic==2.6.4
pymupdf==1.26.4
reportlab==4.1.0
# PageIndex requirements
openai==1.10.0
tiktoken==0.5.2
anthropic==0.18.1
